Fixed Rep Rate on Frequency Scan Panel
Also made spdt width a float instead of int because sometimes we need a decimal place

+++ src/spectrometer_gui/frequency_scan_settings.py
@@ -33,7 +33,7 @@ class DigitizerSettings:
 class TimingSettings:
     rep_rate: int
     valve_mw_delay: int
-    spdt_width: int
+    spdt_width: float
 
 
 @dataclass
